% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r, R/input.R
\docType{data}
\name{input}
\alias{input}
\title{Plik zawirajacy zewnetrzne dane wymagane do dzialania pakietu}
\format{
Szczegolowe opisy danych dstepne sa w ich dokumentacjach.

Ramka danych zawiera:
\describe{
\item{Nat}{Natezenie ruchu (road traffic intensity)}
\item{Segment}{Rodzaj pojazdow (vehicle type)}
\item{Fuel}{Typ paliwa (fuel type)}
\item{Technology}{Technologia silnika (car engine technology)}
}
}
\source{
Obiekt "wskazniki" powstal na podstawie arkusza danych Excel z danymi
pochodzacymi z Europejskiej Agnecji Srodowiska (EEA).
Obiekt "input" zostal wstepnie oobrobiony oraz zapisany do
osobnego obiektu.
}
\usage{
input

input
}
\description{
Obiektem jest losowo wygenerowany wektor.
Pozostale dane sa wybrane z obiektu "wskazniki".
}
\examples{
\dontrun{
 plikdane
}
\dontrun{
plikdane
}
}
\keyword{datasets}
