% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jamfun-package.R
\docType{package}
\name{Jamfun}
\alias{Jamfun}
\title{Jamfun: pakiet służący do obliczania i wizualizacji emisji według EEA}
\description{
Pakiet powstal w ramach przedmiotu realizowanego na studiach
i jest pakietem stworzynym przez studenta. Opiekunem przedmiotu
oraz osoba nadzorujaca prace nad pakietem jest doktor Matusz Rzeszutek.
Zadaniem pakietu jest obliczenie emisji zanieczyszczen, ktore zostaly
wyemitowane do srodowiska w ramach emisji powstalych podczas
funkcjonowania komunikacji (samochodow, pojazdow) we wskazanych
kategoriach pojazdow. W ramach pakietu mozna dane rowniez zwizualizowac.
}
\details{
Pakiet zawiera dwie funkcje:
\code{oblicz_emisje()} obliczanie emisji na podstawie zadanych parametrow
i dostarczonych danych (calculation of emissions)
\code{wizualizuj_emisje()} wizualizacja danych (date visualisation)
}
