% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_emisje.R
\name{oblicz_emisje}
\alias{oblicz_emisje}
\title{Funkcja do obliczanie emisji pochodzacej z transportu drogowego}
\usage{
oblicz_emisje(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("CO", "EC")
)
}
\arguments{
\item{dane}{data.frame - obiekt wejsciowy}

\item{kategoria}{char - kategoria pojazdu}

\item{euro}{char - norma spalin EEA}

\item{mode}{char - tryb jazdy}

\item{substancja}{char - zarejestrowane substancje}
}
\description{
Funkcja oblicza emisje korzystajac z danych zawartych w obiekcie "wskazniki",
a tazke z obiektu "input".
W obiekcie "wskazniki" zapisane sa dane o emisji, ktore zostaly wczesniej
przygotowane i usystematyzowane.
W obiekcie "input" zapisane sa wygenerowane losowo dane o natezenia ruchu.
}
\details{
Wzor wykorzystany do obliczenia emisji:
Emisja = Nat x ((Alpha  x Procent^2 + Beta x Procent + Gamma +
(Delta/Procent)) / (Epsilon x Procent^2 + Zita x Procent + Hta) x (1- Reduction))
}
\examples{
# Obliczenie emisji dla domyslnych parametrow
  tmp <- oblicz_emisje()

}
