% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizuj_emisje.R
\name{wizualizuj_emisje}
\alias{wizualizuj_emisje}
\title{Funkcja do Wizualizacji danych dostarczonych z pakietem}
\usage{
wizualizuj_emisje(
  df = wskazniki,
  kat1 = Category,
  kat2 = "Passenger Cars",
  kategoryczna = Segment,
  numeryczna = Gamma
)
}
\arguments{
\item{df}{data.frame - ramka danych}

\item{kat1}{char - np. kategoria pojazdu}

\item{kat2}{char - np. subkategoria pojazdu}

\item{kategoryczna}{char - zmiena kategoryczna}

\item{numeryczna}{char - parametr szacowania emisji}
}
\description{
Funkcja do Wizualizacji danych dostarczonych z pakietem
}
\examples{
# Wygenerowanie wykresu dla domyslnych danych
wizualizuj_emisje()

}
